pyproject.toml
[project]
name = "coda"
version = "1.0.0"
description = "A codebase agent that learns your code and lives in Slack"
requires-python = ">=3.12,<4"
readme = "README.md"
license-files = ["LICENSE"]
authors = [{ name = "Agno", email = "hello@agno.com" }]

dependencies = [
    "agno",
    "fastapi[standard]",
    "httpx",
    "openai",
    "pgvector",
    "psycopg[binary]",
    "sqlalchemy",
    "opentelemetry-api",
    "opentelemetry-sdk",
    "openinference-instrumentation-agno",
    "pyyaml",
]

[project.optional-dependencies]
dev = ["mypy", "ruff"]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["coda*", "db*", "app*", "tasks*", "evals*"]

[tool.ruff]
line-length = 120
exclude = [".venv*"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401", "F403"]

[tool.mypy]
check_untyped_defs = true
no_implicit_optional = true
warn_unused_configs = true
plugins = ["pydantic.mypy"]
exclude = [".venv*"]

[[tool.mypy.overrides]]
module = ["pgvector.*", "setuptools.*", "agno.*", "httpx.*", "yaml.*"]
ignore_missing_imports = true

[tool.uv.pip]
no-annotate = true
